+++ src/test/java/com/mentionall/cpr2u/education/service/LectureServiceTest.java
@@ -5,10 +5,12 @@ import com.mentionall.cpr2u.user.domain.User;
 import com.mentionall.cpr2u.user.dto.user.SignUpRequestDto;
 import com.mentionall.cpr2u.user.repository.UserRepository;
 import com.mentionall.cpr2u.user.service.AddressService;
+import com.mentionall.cpr2u.user.service.AuthService;
 import com.mentionall.cpr2u.user.service.UserService;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+
 import javax.transaction.Transactional;
 
 import static com.mentionall.cpr2u.education.domain.TestStandard.finalLectureStep;
@@ -20,19 +22,15 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class LectureServiceTest {
 
     @Autowired
-    private LectureService lectureService;
-
-    @Autowired
-    private UserService userService;
-
-    @Autowired
-    private EducationProgressService progressService;
-
+    private AuthService authService;
     @Autowired
     private UserRepository userRepository;
-
+    @Autowired
+    private EducationProgressService progressService;
     @Autowired
     private AddressService addressService;
+    @Autowired
+    private LectureService lectureService;
 
     @BeforeEach
     private void beforeEach() {
@@ -46,7 +44,7 @@ public class LectureServiceTest {
         createLectureCourse();
 
         var address = addressService.readAll().get(0).getGugunList().get(0);
-        userService.signup(new SignUpRequestDto("유저1", "010-1234-1234", address.getId(), "device_token"));
+        authService.signup(new SignUpRequestDto("유저1", "010-1234-1234", address.getId(), "device_token"));
         User user = userRepository.findByPhoneNumber("010-1234-1234").get();
 
         //when
@@ -64,7 +62,7 @@ public class LectureServiceTest {
         createLectureCourse();
 
         var address = addressService.readAll().get(0).getGugunList().get(0);
-        userService.signup(new SignUpRequestDto("유저1", "010-1234-1234", address.getId(), "device_token"));
+        authService.signup(new SignUpRequestDto("유저1", "010-1234-1234", address.getId(), "device_token"));
         User user = userRepository.findByPhoneNumber("010-1234-1234").get();
 
         completeFirstLecture(user);
